Revert all changes to py/torch_tensorrt/fx
Revert "fix: Add automatic type promotion for FX ops"

This reverts commit f1f371663b222e58ec15c335080365f4b2a44a40.

Revert "Moved clamp to impl"

This reverts commit df401dd95ba8f6bbe5777e7c36086b30d1eeea26.

Revert "aten::unsqueeze impl refactor"

This reverts commit b4247358540e353ab1d81252a0000a2131d68528.

Revert "Converter reorg and where operator"

This reverts commit b4da15e9af5e4b92fe2ee76ed6fd71696452bd3b.

Revert "converter reorg and matmul"

This reverts commit 7551eeecfd3c7679cbed048f910bed2e792910c8.

Revert "converter reorg and slice"

This reverts commit 9bbdc9ecf6c8fd04b915dc40567bc67d13c86d38.

Revert "Converter reorg and select operation"

This reverts commit fb70253c1372020c55812f4b22df5f64eb59018a.

Revert "Converter reorg and squeeze operator"

This reverts commit 294545c6361b79656f74f742a84abd6b9b1f1d24.

Revert "Converter reorg and gelu"

This reverts commit 37d11682092d64f1f250581944faddd7c36de176.

Revert "Converter reorg and softmax operation"

This reverts commit 1ba6d139d100b7cc1ba7c95400047ab993e97ce2.

Revert "layer_norm converter"

This reverts commit e0b34b1c35be30e145c2bc484da661c43c23cf9b.

Revert "Converter reorg batch norm"

This reverts commit 59354e5416ee14ca3145b28b389be2ffae2c1821.

Revert "Converter reorg and rsub"

This reverts commit db15d2704f7b12ff4fcee4a267b42bc7c5d37bf6.

Revert "Converter reorg fmod"

This reverts commit ce3fa67655cc0924dc2b046aad8b583c90e283ae.

Revert "Moving elementwise core to impl - rsqrt (FX Converter Refactor [9/N]) <Target: converter_reorg_elementwise> (#1905)"

This reverts commit 7158ca54bd1827e5e0de68107cf19a20238c36f6.

Revert "refactor: Moving elementwise and unary core to impl"

This reverts commit 45e43ca7ccd6ae8fa3d67b124a2a453989d2de47.
