+++ py/torch_tensorrt/fx/converters/acc_ops_converters.py
@@ -27,20 +27,6 @@ from torch_tensorrt.fx.passes.lower_basic_pass import (
 )
 from torch_tensorrt.fx.tracer.acc_tracer.acc_ops import contiguous
 from torch_tensorrt.fx.converters.impl import activation, convolution
-from torch_tensorrt.fx.converters.impl.elementwise import trunc_div
-from torch_tensorrt.fx.converters.impl.elementwise import fmod
-from torch_tensorrt.fx.converters.impl.normalization import batch_norm
-from torch_tensorrt.fx.converters.impl.normalization import layer_norm
-from torch_tensorrt.fx.converters.impl.normalization import softmax
-from torch_tensorrt.fx.converters.impl.elementwise import clamp
-from torch_tensorrt.fx.converters.impl.unary import sign
-from torch_tensorrt.fx.converters.impl.elementwise.base import (
-    convert_binary_elementwise,
-)
-from torch_tensorrt.fx.converters.impl.unary.base import convert_unary
-from torch_tensorrt.fx.converters.impl.shape import get_shape_with_dynamic_shape
-from torch_tensorrt.fx.converters.impl.squeeze import squeeze
-from torch_tensorrt.fx.converters.impl.unsqueeze import unsqueeze
 
 _LOGGER: logging.Logger = logging.getLogger(__name__)
 
@@ -92,14 +78,13 @@ def trt_transposed_linear_converter(network, target, args, kwargs, name):
         trt.MatrixOperation.NONE,
     )
     set_layer_name(layer, target, f"{name}_mm")
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.TORCHTRT_LOWERED,
-        f"{name}_add",
-        trt.ElementWiseOperation.SUM,
         layer.get_output(0),
         bias,
+        trt.ElementWiseOperation.SUM,
+        target,
+        f"{name}_add",
     )
 
 
@@ -527,34 +512,216 @@ def acc_ops_batch_norm(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return batch_norm(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        kwargs["input"],
-        kwargs["weight"],
-        kwargs["bias"],
-        kwargs["running_mean"],
-        kwargs["running_var"],
-        kwargs["training"],
-        kwargs["momentum"],
-        kwargs["eps"],
+    input_val = kwargs["input"]
+
+    if not isinstance(input_val, TRTTensor):
+        raise RuntimeError(
+            f"BatchNorm2d received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+
+    if has_dynamic_shape(input_val.shape):
+        assert input_val.shape[1] != -1, "Channel dim can't be dynamic for batch norm."
+
+    scale = cast(
+        torch.Tensor, to_numpy(cast(torch.Tensor, kwargs["weight"]))
+    ) / np.sqrt(
+        cast(torch.Tensor, to_numpy(cast(torch.Tensor, kwargs["running_var"])))
+        + cast(float, kwargs["eps"])
     )
+
+    bias = (
+        to_numpy(cast(torch.Tensor, kwargs["bias"]))
+        - to_numpy(cast(torch.Tensor, kwargs["running_mean"])) * scale
+    )
+    power = np.ones_like(scale)
+
+    # For BatchNorm1d, reshape 1d to 2d
+    output_shape = input_val.shape
+    if not network.has_implicit_batch_dimension and len(input_val.shape) < 4:
+        assert (
+            len(get_dynamic_dims(input_val.shape)) <= 1
+        ), "BatchNorm1D with more than one dynamic dims is not currently supported."
+        reshape_layer = network.add_shuffle(input_val)
+        if len(input_val.shape) == 2:
+            reshape_layer.reshape_dims = (input_val.shape[0], input_val.shape[1], 1, 1)
+        else:  # len(input_val.shape) == 3
+            reshape_layer.reshape_dims = (
+                input_val.shape[0],
+                input_val.shape[1],
+                input_val.shape[2],
+                1,
+            )
+        set_layer_name(reshape_layer, target, f"{name}_reshape_2d")
+        input_val = reshape_layer.get_output(0)
+    layer = network.add_scale(input_val, trt.ScaleMode.CHANNEL, bias, scale, power)
+    set_layer_name(layer, target, name)
+
+    # For BatchNorm1d, reshape output back to 1d
+    if not network.has_implicit_batch_dimension and len(output_shape) < 4:
+        reshape_output_layer = network.add_shuffle(layer.get_output(0))
+        reshape_output_layer.reshape_dims = tuple(output_shape)
+        set_layer_name(reshape_output_layer, target, f"{name}_reshape_1d")
+        layer = reshape_output_layer
+    return layer.get_output(0)
 
 
 @tensorrt_converter(acc_ops.layer_norm)
 def acc_ops_layer_norm(network, target, args, kwargs, name):
-    return layer_norm(
+    input_val = kwargs["input"]
+
+    if not isinstance(input_val, trt.tensorrt.ITensor):
+        raise RuntimeError(
+            f"LayerNorm received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+
+    gamma = kwargs["weight"].detach().cpu().float().numpy()
+    gamma_field = trt.PluginField("gamma", gamma, trt.PluginFieldType.FLOAT32)
+    beta = kwargs["bias"].detach().cpu().float().numpy()
+    beta_field = trt.PluginField("beta", beta, trt.PluginFieldType.FLOAT32)
+    eps_field = trt.PluginField(
+        "eps", np.array([kwargs["eps"]], dtype=np.float32), trt.PluginFieldType.FLOAT32
+    )
+    normalized_shape = kwargs["normalized_shape"]
+    try:
+        normalized_shape = np.array(normalized_shape, dtype=np.int32)
+    except TypeError:
+        _LOGGER.error(
+            f"Unable to convert normalized_shape with value {normalized_shape} to a field, fall back to []"
+        )
+        normalized_shape = np.array([], dtype=np.int32)
+
+    normalized_shape_filed = trt.PluginField(
+        "normalized_shape", normalized_shape, trt.PluginFieldType.INT32
+    )
+    field_collection = trt.PluginFieldCollection(
+        [gamma_field, beta_field, eps_field, normalized_shape_filed]
+    )
+
+    try:
+        if network.has_implicit_batch_dimension:
+            plugin = get_trt_plugin("layer_norm", field_collection, "1", "fx2trt")
+        else:
+            plugin = get_trt_plugin("LayerNormDynamic", field_collection, "1", "fx2trt")
+    except AssertionError:
+        _LOGGER.error(
+            "Unable to find layer norm plugin, fall back to TensorRT implementation."
+        )
+        return layer_norm(network, target, args, kwargs, name)
+    layer = network.add_plugin_v2([input_val], plugin)
+    layer.name = name
+    return layer.get_output(0)
+
+
+def layer_norm(
+    network: TRTNetwork,
+    target: Target,
+    args: Tuple[Argument, ...],
+    kwargs: Dict[str, Argument],
+    name: str,
+) -> Union[TRTTensor, Sequence[TRTTensor]]:
+    input_val = kwargs["input"]
+
+    if not isinstance(input_val, TRTTensor):
+        raise RuntimeError(
+            f"LayerNorm received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+
+    shape = kwargs["weight"].shape  # type: ignore[union-attr]
+    broadcasted_shape = (1,) * (len(input_val.shape) - len(shape)) + shape
+    gamma = to_numpy(kwargs["weight"].reshape(*shape))  # type: ignore[union-attr]
+    beta = to_numpy(kwargs["bias"].reshape(*shape))  # type: ignore[union-attr]
+    eps = kwargs["eps"]
+
+    axes = 0
+    for d in range(len(shape)):
+        axes |= 1 << (len(input_val.shape) - d - 1)
+
+    # E[x]
+    mean_expected_layer = network.add_reduce(
+        input_val, trt.ReduceOperation.AVG, axes, keep_dims=True
+    )
+    set_layer_name(mean_expected_layer, target, f"{name}_mean_expected")
+
+    # X-E[x]
+    sub_trt = add_binary_elementwise_layer(
         network,
+        input_val,
+        mean_expected_layer.get_output(0),
+        trt.ElementWiseOperation.SUB,
         target,
-        SourceIR.ACC,
+        f"{name}_sub",
+    )
+    # Variance = mean(pow(x_sub_mean,2))
+    pow_tensor = network.add_constant(
+        (1,) * len(input_val.shape),
+        trt.Weights(np.ascontiguousarray([2.0], dtype=np.float32)),
+    )
+    pow_tensor.name = f"{name}_power"
+    pow_var = add_binary_elementwise_layer(
+        network,
+        sub_trt,
+        pow_tensor.get_output(0),
+        trt.ElementWiseOperation.POW,
+        target,
+        f"{name}_pow_var",
+    )
+    mean_trt_layer = network.add_reduce(
+        pow_var, trt.ReduceOperation.AVG, axes, keep_dims=True
+    )
+    set_layer_name(mean_trt_layer, target, f"{name}_mean")
+    # Variance + eps
+    eps_tensor = network.add_constant(
+        (1,) * len(input_val.shape),
+        trt.Weights(np.ascontiguousarray([eps], dtype=np.float32)),
+    )
+    eps_tensor.name = f"{name}_eps"
+    add_trt = add_binary_elementwise_layer(
+        network,
+        mean_trt_layer.get_output(0),
+        eps_tensor.get_output(0),
+        trt.ElementWiseOperation.SUM,
+        target,
+        f"{name}_add",
+    )
+    # SQRT((Var + eps))
+    sqrt_trt = add_unary_layer(
+        network, add_trt, trt.UnaryOperation.SQRT, target, f"{name}_sqrt"
+    )
+    # (x - E[x]) / sqrt((var + eps))
+    div_trt = add_binary_elementwise_layer(
+        network,
+        sub_trt,
+        sqrt_trt,
+        trt.ElementWiseOperation.DIV,
+        target,
+        f"{name}_div_trt",
+    )
+
+    assert gamma is not None
+    gamma_tensor = network.add_constant(gamma.shape, trt.Weights(np.ascontiguousarray(gamma)))  # type: ignore[attr-defined]
+    gamma_tensor.name = f"{name}_gamma"
+    assert beta is not None
+    beta_tensor = network.add_constant(gamma.shape, trt.Weights(np.ascontiguousarray(beta)))  # type: ignore[attr-defined]
+    beta_tensor.name = f"{name}_beta"
+    # y * gamma + beta
+    scale_layer = add_binary_elementwise_layer(
+        network,
+        div_trt,
+        gamma_tensor.get_output(0),
+        trt.ElementWiseOperation.PROD,
+        target,
+        f"{name}_scale",
+    )
+    return add_binary_elementwise_layer(
+        network,
+        scale_layer,
+        beta_tensor.get_output(0),
+        trt.ElementWiseOperation.SUM,
+        target,
         name,
-        kwargs["input"],
-        kwargs["normalized_shape"],
-        kwargs["weight"],
-        kwargs["bias"],
-        kwargs["eps"],
     )
 
 
@@ -566,7 +733,37 @@ def acc_ops_softmax(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return softmax(network, target, SourceIR.ACC, name, kwargs["input"], kwargs["dim"])
+    input_val = kwargs["input"]
+    input_ranks = len(input_val.shape) + (1 if network.has_implicit_batch_dimension else 0)  # type: ignore[union-attr]
+
+    if not isinstance(input_val, TRTTensor):
+        raise RuntimeError(
+            f"softmax received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+
+    # Used to get dim when dim is None. Copied from PyTorch softmax implementation.
+    def get_softmax_dim(ndim: int) -> int:
+        if ndim == 0 or ndim == 1 or ndim == 3:
+            ret = 0
+        else:
+            ret = 1
+        return ret
+
+    if kwargs["dim"] is None:
+        dim = get_softmax_dim(input_ranks)
+    else:
+        dim = cast(int, kwargs["dim"])
+
+    dim = get_positive_dim(dim, input_ranks)
+    if network.has_implicit_batch_dimension:
+        assert dim != 0, "Can't apply softmax on batch dimension when it's implicit."
+        dim -= 1
+
+    layer = network.add_softmax(input_val)
+    layer.axes = 1 << dim
+    set_layer_name(layer, target, name)
+    return layer.get_output(0)
 
 
 @tensorrt_converter(acc_ops.tile)
@@ -627,14 +824,13 @@ def acc_ops_tile(
             else:
                 d = get_trt_tensor(network, d, f"{name}_{i}")
             shape.append(d)
-            mul = convert_binary_elementwise(
+            mul = add_binary_elementwise_layer(
                 network,
-                target,
-                SourceIR.ACC,
-                f"{name}_mul_{i}",
-                trt.ElementWiseOperation.PROD,
                 s,
                 d,
+                trt.ElementWiseOperation.PROD,
+                target,
+                f"{name}_mul_{i}",
             )
             shapes.append(mul)
         dims = shape
@@ -663,14 +859,13 @@ def acc_ops_tile(
             dims_tensor = concat_dims_layer.get_output(0)
         input_shape_layer = network.add_shape(input_val)
         input_shape_layer.name = f"{name}_slice_input_shape"
-        slice_shapes_tensor = convert_binary_elementwise(
+        slice_shapes_tensor = add_binary_elementwise_layer(
             network,
-            target,
-            SourceIR.ACC,
-            f"{name}_slice_shapes",
-            trt.ElementWiseOperation.PROD,
             input_shape_layer.get_output(0),
             dims_tensor,
+            trt.ElementWiseOperation.PROD,
+            target,
+            f"{name}_slice_shapes",
         )
         layer.set_input(1, starts_tensor)
         layer.set_input(2, slice_shapes_tensor)
@@ -691,22 +886,9 @@ def acc_ops_sign(
     if trt.__version__ >= "8.2" and not network.has_implicit_batch_dimension:
         input_val = kwargs["input"]
         operation_type = trt.UnaryOperation.SIGN
-        return convert_unary(
-            network,
-            target,
-            SourceIR.ACC,
-            name,
-            operation_type,
-            input_val,
-        )
+        return add_unary_layer(network, input_val, operation_type, target, name)
 
-    return sign(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        input_val,
-    )
+    return sign(network, input_val, target, name)
 
 
 @tensorrt_converter(acc_ops.relu)
@@ -808,14 +990,7 @@ def acc_ops_sin(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.SIN
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.cos)
@@ -828,14 +1003,7 @@ def acc_ops_cos(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.COS
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.tan)
@@ -848,14 +1016,7 @@ def acc_ops_tan(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.TAN
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.sinh)
@@ -868,14 +1029,7 @@ def acc_ops_sinh(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.SINH
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.cosh)
@@ -888,14 +1042,7 @@ def acc_ops_cosh(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.COSH
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.tanh)
@@ -925,14 +1072,7 @@ def acc_ops_asin(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.ASIN
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.acos)
@@ -945,14 +1085,7 @@ def acc_ops_acos(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.ACOS
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.atan)
@@ -965,14 +1098,7 @@ def acc_ops_atan(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.ATAN
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.exp)
@@ -985,14 +1111,7 @@ def acc_ops_exp(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.EXP
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.log)
@@ -1005,14 +1124,7 @@ def acc_ops_log(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.LOG
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.sqrt)
@@ -1025,14 +1137,7 @@ def acc_ops_sqrt(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.SQRT
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.reciprocal)
@@ -1045,14 +1150,7 @@ def acc_ops_reciprocal(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.RECIP
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.abs)
@@ -1065,14 +1163,7 @@ def acc_ops_abs(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.ABS
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.neg)
@@ -1085,14 +1176,7 @@ def acc_ops_neg(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.NEG
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.floor)
@@ -1105,14 +1189,7 @@ def acc_ops_floor(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.FLOOR
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.ceil)
@@ -1125,14 +1202,7 @@ def acc_ops_ceil(
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
     input_val = kwargs["input"]
     operation_type = trt.UnaryOperation.CEIL
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.sum)
@@ -1307,14 +1377,13 @@ def acc_ops_maximum(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.MAX,
         kwargs["input"],
         kwargs["other"],
+        trt.ElementWiseOperation.MAX,
+        target,
+        name,
     )
 
 
@@ -1326,14 +1395,13 @@ def acc_ops_minimum(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.MIN,
         kwargs["input"],
         kwargs["other"],
+        trt.ElementWiseOperation.MIN,
+        target,
+        name,
     )
 
 
@@ -1392,14 +1460,7 @@ def acc_ops_logical_not(
     # cast to bool type
     if input_val.dtype in (trt.float32, trt.float16, trt.int32):
         input_val = type_cast(network, target, f"{name}_input", input_val, trt.bool)
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        operation_type,
-        input_val,
-    )
+    return add_unary_layer(network, input_val, operation_type, target, name)
 
 
 @tensorrt_converter(acc_ops.logical_and, no_implicit_batch_dim=True)
@@ -1445,14 +1506,8 @@ def acc_ops_logical_and(
         input_t = type_cast(network, target, f"{name}_input", input_t, trt.bool)
     if other_t.dtype != trt.bool:
         other_t = type_cast(network, target, f"{name}_other", other_t, trt.bool)
-    return convert_binary_elementwise(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.AND,
-        input_t,
-        other_t,
+    return add_binary_elementwise_layer(
+        network, input_t, other_t, trt.ElementWiseOperation.AND, target, name
     )
 
 
@@ -1476,24 +1531,11 @@ def acc_ops_ne(
     other_t = get_trt_tensor(network, other_t, f"{name}_other_t")
 
     input_t, other_t = dtype_uniform(network, target, name, input_t, other_t)
-    eq_t = convert_binary_elementwise(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.EQUAL,
-        input_t,
-        other_t,
+    eq_t = add_binary_elementwise_layer(
+        network, input_t, other_t, trt.ElementWiseOperation.EQUAL, target, name
     )
 
-    return convert_unary(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.UnaryOperation.NOT,
-        eq_t,
-    )
+    return add_unary_layer(network, eq_t, trt.UnaryOperation.NOT, target, name)
 
 
 @tensorrt_converter(acc_ops.eq, no_implicit_batch_dim=True)
@@ -1516,14 +1558,8 @@ def acc_ops_eq(
     other_t = get_trt_tensor(network, other_t, f"{name}_other_t")
 
     input_t, other_t = dtype_uniform(network, target, name, input_t, other_t)
-    return convert_binary_elementwise(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.EQUAL,
-        input_t,
-        other_t,
+    return add_binary_elementwise_layer(
+        network, input_t, other_t, trt.ElementWiseOperation.EQUAL, target, name
     )
 
 
@@ -1547,14 +1583,8 @@ def acc_ops_gt(
     other_t = get_trt_tensor(network, other_t, f"{name}_other_t")
 
     input_t, other_t = dtype_uniform(network, target, name, input_t, other_t)
-    return convert_binary_elementwise(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.GREATER,
-        input_t,
-        other_t,
+    return add_binary_elementwise_layer(
+        network, input_t, other_t, trt.ElementWiseOperation.GREATER, target, name
     )
 
 
@@ -1578,14 +1608,8 @@ def acc_ops_lt(
     other_t = get_trt_tensor(network, other_t, f"{name}_other_t")
 
     input_t, other_t = dtype_uniform(network, target, name, input_t, other_t)
-    return convert_binary_elementwise(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.LESS,
-        input_t,
-        other_t,
+    return add_binary_elementwise_layer(
+        network, input_t, other_t, trt.ElementWiseOperation.LESS, target, name
     )
 
 
@@ -1621,14 +1645,8 @@ def acc_ops_logical_or(
         set_layer_name(layer_o, target, f"{name}_other_dtype_change")
         other_t = layer_o.get_output(0)
 
-    return convert_binary_elementwise(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.OR,
-        input_t,
-        other_t,
+    return add_binary_elementwise_layer(
+        network, input_t, other_t, trt.ElementWiseOperation.OR, target, name
     )
 
 
@@ -1664,14 +1682,8 @@ def acc_ops_logical_xor(
         set_layer_name(layer_o, target, f"{name}_other_dtype_change")
         other_t = layer_o.get_output(0)
 
-    return convert_binary_elementwise(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.XOR,
-        input_t,
-        other_t,
+    return add_binary_elementwise_layer(
+        network, input_t, other_t, trt.ElementWiseOperation.XOR, target, name
     )
 
 
@@ -1766,14 +1778,27 @@ def acc_ops_fmod(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return fmod(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        kwargs["input"],
-        kwargs["other"],
+    # NOTE: TRT doesnt currently implement fmod so we need multiple operations to perform it
+    trunc_div_value = trunc_div(
+        kwargs["input"], kwargs["other"], network, target, name + "_trunc_div"
     )
+    prod_value = add_binary_elementwise_layer(
+        network,
+        trunc_div_value,
+        kwargs["other"],
+        trt.ElementWiseOperation.PROD,
+        target,
+        name + "_prod",
+    )
+    sub_value = add_binary_elementwise_layer(
+        network,
+        kwargs["input"],
+        prod_value,
+        trt.ElementWiseOperation.SUB,
+        target,
+        name + "_sub",
+    )
+    return sub_value
 
 
 # T113156424 embedding implemenatation is very limited and shows no usage in hf models due to the indices are int64.
@@ -1958,14 +1983,40 @@ def acc_ops_squeeze(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return squeeze(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        kwargs["input"],
-        kwargs["dim"],
+    input_val = kwargs["input"]
+
+    if not isinstance(input_val, TRTTensor):
+        raise RuntimeError(
+            f"squeeze received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+
+    dim = cast(Optional[int], kwargs["dim"] if "dim" in kwargs else None)
+    # Squeeze with dim=None would only work in explicit batch dim mode without any dynamic
+    # dim, which is a very rare case. For now we just claim not supporting dim=None.
+    assert dim is not None, "We don't support dim=None right now for squeeze."
+
+    dim = get_positive_dim(
+        dim, len(input_val.shape) + (1 if network.has_implicit_batch_dimension else 0)
     )
+    if network.has_implicit_batch_dimension:
+        assert dim != 0, "We don't support squeeze batch dim when it's implicit."
+        dim -= 1
+
+    assert input_val.shape[dim] != -1, "We don't support squeeze dynamic dim."
+    assert (
+        len(get_dynamic_dims(input_val.shape)) <= 1
+    ), "Currently more than one dynamic dim for input to squeeze is not supported."
+
+    output_shape = []
+    for i, s in enumerate(input_val.shape):
+        if i == dim and s == 1:
+            continue
+        output_shape.append(s)
+    layer = network.add_shuffle(input_val)
+    layer.reshape_dims = tuple(output_shape)
+    set_layer_name(layer, target, name)
+    return layer.get_output(0)
 
 
 @tensorrt_converter(acc_ops.add)
@@ -1976,14 +2027,13 @@ def acc_ops_add(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.SUM,
         kwargs["input"],
         kwargs["other"],
+        trt.ElementWiseOperation.SUM,
+        target,
+        name,
     )
 
 
@@ -1995,14 +2045,13 @@ def acc_ops_sub(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.SUB,
         kwargs["input"],
         kwargs["other"],
+        trt.ElementWiseOperation.SUB,
+        target,
+        name,
     )
 
 
@@ -2014,14 +2063,13 @@ def acc_ops_div(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.DIV,
         kwargs["input"],
         kwargs["other"],
+        trt.ElementWiseOperation.DIV,
+        target,
+        name,
     )
 
 
@@ -2033,14 +2081,13 @@ def acc_ops_floor_div(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.FLOOR_DIV,
         kwargs["input"],
         kwargs["other"],
+        trt.ElementWiseOperation.FLOOR_DIV,
+        target,
+        name,
     )
 
 
@@ -2052,14 +2099,7 @@ def acc_ops_trunc_div(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return trunc_div(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        kwargs["input"],
-        kwargs["other"],
-    )
+    return trunc_div(kwargs["input"], kwargs["other"], network, target, name)
 
 
 @tensorrt_converter(acc_ops.mul)
@@ -2070,14 +2110,13 @@ def acc_ops_mul(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.PROD,
         kwargs["input"],
         kwargs["other"],
+        trt.ElementWiseOperation.PROD,
+        target,
+        name,
     )
 
 
@@ -2089,14 +2128,13 @@ def acc_ops_pow(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return convert_binary_elementwise(
+    return add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        name,
-        trt.ElementWiseOperation.POW,
         kwargs["input"],
         kwargs["exponent"],
+        trt.ElementWiseOperation.POW,
+        target,
+        name,
     )
 
 
@@ -2108,14 +2146,36 @@ def acc_ops_unsqueeze(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return unsqueeze(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        input_t=kwargs["input"],
-        dim=kwargs["dim"],
+    input_t = kwargs["input"]
+    input_val = get_trt_tensor(network, input_t, f"{name}_input_t")
+    if not isinstance(input_val, TRTTensor):
+        raise RuntimeError(
+            f"unsqueeze received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+
+    dim = cast(int, kwargs["dim"])
+    input_shape = input_val.shape
+    input_shape_size = (
+        len(input_val.shape) + 1
+        if network.has_implicit_batch_dimension
+        else len(input_val.shape)
     )
+    dim = get_positive_dim(dim, input_shape_size + 1)
+
+    if network.has_implicit_batch_dimension:
+        assert dim != 0
+        dim -= 1
+
+    assert (
+        len(get_dynamic_dims(input_val.shape)) <= 1
+    ), "Currently we don't support unsqueeze with more than one dynamic dims."
+    layer = network.add_shuffle(input_val)
+    layer.reshape_dims = (
+        tuple(input_val.shape)[:dim] + (1,) + tuple(input_val.shape)[dim:]
+    )
+    set_layer_name(layer, target, name)
+    return layer.get_output(0)
 
 
 @tensorrt_converter(acc_ops.topk)
@@ -2380,12 +2440,7 @@ def acc_ops_slice_tensor(
 
     if dynamic_shape > 0:
         output_shape = get_shape_with_dynamic_shape(
-            network,
-            target,
-            SourceIR.ACC,
-            name,
-            output_shape,
-            input_val,
+            network, output_shape, input_val, target, name
         )
     layer = network.add_slice(
         input_val,
@@ -2629,12 +2684,7 @@ def acc_ops_split(
         start[dim] = offset
         if dynamic_shape:
             shape = get_shape_with_dynamic_shape(
-                network,
-                target,
-                SourceIR.ACC,
-                f"{name}_shape_{i}",
-                shape,
-                input_val,
+                network, shape, input_val, target, f"{name}_shape_{i}"
             )
         layer = network.add_slice(
             input_val, start=start, shape=[] if dynamic_shape else shape, stride=stride
@@ -2698,16 +2748,47 @@ def acc_ops_linear(
 
     if kwargs["bias"] is not None:
         bias = get_trt_tensor(network, kwargs["bias"], f"{name}_bias")  # type: ignore[arg-type]
-        res = convert_binary_elementwise(
+        res = add_binary_elementwise_layer(
             network,
-            target,
-            SourceIR.ACC,
-            f"{name}_add",
-            trt.ElementWiseOperation.SUM,
             matmul_layer.get_output(0),
             bias,
+            trt.ElementWiseOperation.SUM,
+            target,
+            f"{name}_add",
         )
     return res
+
+
+def add_clamp(network, input, val, op, name):
+    if not len(input.shape):
+        # clamping scalar
+        acc_ops_clamp_trt = get_trt_tensor(
+            network,
+            squeeze_left(
+                torch.tensor(
+                    [val], dtype=unified_dtype_converter(input.dtype, Frameworks.TORCH)
+                )
+            ),
+            f"{name}_clamp_{val}",
+        )
+    else:
+        acc_ops_clamp_shape = (1,) * len(input.shape)  # broadcast all dimensions
+        acc_ops_clamp_tensor = (
+            (
+                val
+                * torch.ones(
+                    acc_ops_clamp_shape,
+                    dtype=unified_dtype_converter(input.dtype, Frameworks.TORCH),
+                )
+            )
+            .cpu()
+            .numpy()
+        )
+        acc_ops_clamp_trt = network.add_constant(
+            acc_ops_clamp_shape, acc_ops_clamp_tensor
+        ).get_output(0)
+    layer = network.add_elementwise(input, acc_ops_clamp_trt, op)
+    return layer
 
 
 @tensorrt_converter(acc_ops.clamp)
@@ -2718,15 +2799,30 @@ def acc_ops_clamp(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return clamp.clamp(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        input_val=kwargs["input"],
-        min_val=kwargs["min"],
-        max_val=kwargs["max"],
-    )
+    input_val = kwargs["input"]
+    min_val = kwargs["min"]
+    max_val = kwargs["max"]
+
+    if not isinstance(input_val, TRTTensor):
+        raise RuntimeError(
+            f"Clamp received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+
+    if min_val is not None:
+        clamp_min_layer = add_clamp(
+            network, input_val, min_val, trt.ElementWiseOperation.MAX, name
+        )
+        set_layer_name(clamp_min_layer, target, f"{name}_clamp_min")
+        input_val = clamp_min_layer.get_output(0)
+    if max_val is not None:
+        clamp_max_layer = add_clamp(
+            network, input_val, max_val, trt.ElementWiseOperation.MIN, name
+        )
+        set_layer_name(clamp_max_layer, target, f"{name}_clamp_max")
+        input_val = clamp_max_layer.get_output(0)
+
+    return input_val
 
 
 @tensorrt_converter(acc_ops.tuple_construct)
@@ -2849,14 +2945,7 @@ def acc_ops_getitem(
         i += 1
 
     if dynamic_shape:
-        size = get_shape_with_dynamic_shape(
-            network,
-            target,
-            SourceIR.ACC,
-            name,
-            size,
-            input_val,
-        )
+        size = get_shape_with_dynamic_shape(network, size, input_val, target, name)
 
     layer = network.add_slice(
         input=input_val,
@@ -3200,14 +3289,33 @@ def acc_ops_gelu(
     kwargs: Dict[str, Argument],
     name: str,
 ) -> Union[TRTTensor, Sequence[TRTTensor]]:
-    return activation.gelu(
-        network,
-        target,
-        SourceIR.ACC,
-        name,
-        kwargs["input"],
-        kwargs["approximate"],
+    input_val = kwargs["input"]
+    approximate = kwargs["approximate"]
+    if approximate != "none":
+        raise RuntimeError("GeLU converter currently doesn't support fast gelu compute")
+    if not isinstance(input_val, TRTTensor):
+        raise RuntimeError(
+            f"GELU received input {input_val} that is not part "
+            "of the TensorRT region!"
+        )
+    if network.has_implicit_batch_dimension:
+        raise RuntimeError(
+            "GeLU converter currently doesn't support implicit batch dimension"
+        )
+
+    plugin_name = "CustomGeluPluginDynamic"
+    # type_id 0 for float32, 1 for  float16
+    type_id = trt.PluginField(
+        "type_id", np.array(0, dtype=np.int32), trt.PluginFieldType.INT32
     )
+    field_collection = TRTPluginFieldCollection([type_id])
+    plugin_version = "1"
+
+    plugin = get_trt_plugin(plugin_name, field_collection, plugin_version)
+
+    layer = network.add_plugin_v2([input_val], plugin)
+    set_layer_name(layer, target, name)
+    return layer.get_output(0)
 
 
 @tensorrt_converter(acc_ops.chunk)
@@ -3262,12 +3370,7 @@ def acc_ops_chunk(
         shape[dim] = min(split_size, max_offset - offset)
         if dynamic_shape:
             shape = get_shape_with_dynamic_shape(
-                network,
-                target,
-                SourceIR.ACC,
-                f"{name}_{i}",
-                shape,
-                input_val,
+                network, shape, input_val, target, f"{name}_{i}"
             )
         start[dim] = offset
         layer = network.add_slice(
@@ -3332,14 +3435,13 @@ def acc_ops_cumsum(
     set_layer_name(running_sum, target, f"{name}_running_sum_1")
     running_sum_tensor = running_sum.get_output(0)
 
-    current_sum = convert_binary_elementwise(
+    current_sum = add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        f"{name}_sum_1",
-        trt.ElementWiseOperation.SUM,
         data,
         running_sum_tensor,
+        trt.ElementWiseOperation.SUM,
+        target,
+        f"{name}_sum_1",
     )
     running_sum.set_input(1, current_sum)
 
@@ -3347,14 +3449,13 @@ def acc_ops_cumsum(
     set_layer_name(running_sum, target, f"{name}_running_sum_2")
     running_sum_tensor = running_sum.get_output(0)
 
-    current_sum = convert_binary_elementwise(
+    current_sum = add_binary_elementwise_layer(
         network,
-        target,
-        SourceIR.ACC,
-        f"{name}_sum_2",
-        trt.ElementWiseOperation.SUM,
         data,
         running_sum_tensor,
+        trt.ElementWiseOperation.SUM,
+        target,
+        f"{name}_sum_2",
     )
     running_sum.set_input(1, current_sum)
