--- py/torch_tensorrt/fx/converters/impl/elementwise/__init__.py
@@ -1,1 +1,0 @@
-from .ops import *
