--- py/torch_tensorrt/fx/converters/impl/unary/__init__.py
@@ -1,1 +1,0 @@
-from .ops import *
