--- py/torch_tensorrt/fx/converters/impl/unsqueeze.py
@@ -1,63 +1,0 @@
-import operator
-import warnings
-from typing import Optional, cast, Any
-
-import numpy as np
-
-import tensorrt as trt
-import torch
-from torch.fx.node import Target
-
-from torch_tensorrt.fx.types import TRTNetwork, TRTTensor, Shape
-from torch_tensorrt.fx.converters.converter_utils import (
-    SourceIR,
-    get_positive_dim,
-    get_trt_tensor,
-)
-
-from torch_tensorrt.fx.converters.converter_utils import (
-    SourceIR,
-    get_positive_dim,
-    set_layer_name,
-)
-
-from torch_tensorrt.fx.utils import get_dynamic_dims
-
-
-def unsqueeze(
-    network: TRTNetwork,
-    target: Target,
-    source_ir: Optional[SourceIR],
-    name: str,
-    input_t,
-    dim,
-) -> TRTTensor:
-    input_val = get_trt_tensor(network, input_t, f"{name}_input_t")
-    if not isinstance(input_val, TRTTensor):
-        raise RuntimeError(
-            f"unsqueeze received input {input_val} that is not part "
-            "of the TensorRT region!"
-        )
-
-    dim = cast(int, dim)
-    input_shape = input_val.shape
-    input_shape_size = (
-        len(input_val.shape) + 1
-        if network.has_implicit_batch_dimension
-        else len(input_val.shape)
-    )
-    dim = get_positive_dim(dim, input_shape_size + 1)
-
-    if network.has_implicit_batch_dimension:
-        assert dim != 0
-        dim -= 1
-
-    assert (
-        len(get_dynamic_dims(input_val.shape)) <= 1
-    ), "Currently we don't support unsqueeze with more than one dynamic dims."
-    layer = network.add_shuffle(input_val)
-    layer.reshape_dims = (
-        tuple(input_val.shape)[:dim] + (1,) + tuple(input_val.shape)[dim:]
-    )
-    set_layer_name(layer, target, name)
-    return layer.get_output(0)
