+++ py/torch_tensorrt/fx/test/converters/acc_op/test_squeeze.py
@@ -12,12 +12,7 @@ class TestSqueeze(AccTestCase):
                 return x.squeeze(2)
 
         inputs = [torch.randn(1, 2, 1)]
-        self.run_test(
-            Squeeze(),
-            inputs,
-            expected_ops={acc_ops.squeeze},
-            test_implicit_batch_dim=False,
-        )
+        self.run_test(Squeeze(), inputs, expected_ops={acc_ops.squeeze})
 
     # Testing with shape=(-1, -1, -1, -1) results in error:
     # AssertionError: We don't support squeeze dynamic dim.
@@ -38,9 +33,7 @@ class TestSqueeze(AccTestCase):
             ),
         ]
         self.run_test_with_dynamic_shape(
-            Squeeze(),
-            input_specs,
-            expected_ops={acc_ops.squeeze},
+            Squeeze(), input_specs, expected_ops={acc_ops.squeeze}
         )
