--- py/torch_tensorrt/fx/test/converters/aten_op/test_where_aten.py
@@ -1,33 +1,0 @@
-import torch
-import torch.nn as nn
-from parameterized import parameterized
-from torch.testing._internal.common_utils import run_tests
-from torch_tensorrt.fx.tools.common_fx2trt import DispatchTestCase, InputTensorSpec
-
-
-class TestWhereConverter(DispatchTestCase):
-    @parameterized.expand(
-        [
-            ("2d_condition_xshape_yshape", (2, 2), (2, 2)),
-            ("2d_broadcast_condition_xshape_yshape", (2, 2), (2, 1)),
-            ("3d_condition_xshape_yshape", (2, 2, 1), (2, 2, 1)),
-            ("2d_3d_condition_xshape_yshape", (2, 2), (1, 2, 2)),
-        ]
-    )
-    def test_(self, _, x_size, y_size):
-        class Where(nn.Module):
-            def forward(self, condition, x, y):
-                return torch.where(condition, x, y)
-
-        inputX = torch.randn(*x_size)
-        inputOther = torch.randn(*y_size)
-        condition = inputX < 0
-        self.run_test(
-            Where(),
-            (condition, inputX, inputOther),
-            expected_ops={torch.ops.aten.where.self},
-        )
-
-
-if __name__ == "__main__":
-    run_tests()
